move methods and protocols to utils.py
